Backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
gunicorn
django-cors-headers
python-dotenv
pillow
